feat: Add clearFilter method to TopCarousel
The clearFilter method is added to the TopCarousel widget to reset the filter and fetch all restaurants. This method sets the gettingRestaurants flag to true and calls the getLocal method to fetch the restaurants based on the user's location.

This commit enhances the functionality of the TopCarousel widget by allowing users to clear the applied filter and view all restaurants.